backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
requests
djangorestframework-simplejwt
google-cloud-vision
google-generativeai